Changes made for better functioning of the shell

+++ Userland/SampleCodeModule/include/clock.h
@@ -7,7 +7,7 @@ typedef struct clock{
     uint8_t hours;
     uint8_t dayOfMonth;
     uint8_t month;
-    uint8_t year;
+    uint16_t year;
 }clock;
 
 void time();
